evering::uring

Struct UringB

pub struct UringB<A, B, Ext = ()>(/* private fields */);

Implementations

impl<A, B, Ext> UringB<A, B, Ext>

pub fn into_raw(self) -> RawUring<A, B, Ext>

pub fn dispose_raw(self) -> Result<RawUring<A, B, Ext>, DisposeError>

Drops this Uring and all enqueued entries.

It does nothing and returns an error if self is still connected. Otherwise, the returned RawUring is safe to deallocate without synchronization.

pub unsafe fn from_raw(uring: RawUring<A, B, Ext>) -> Self

Safety

The specified RawUring must be a valid value returned from into_raw.